added a few resources
I'm not sure if this is allowed, but I've added a few resources I wish I knew I had when I began learning. 

Rails Tutorial by Michael Hartl is hailed the holy grail of the rails tutorials, and is updated with every new version release. 

CS50x is a course by HARVARD, and covers a lot of coding fundamentals in a really fun way (we should really consider starting study groups for this course). The used to teach code using C, but have since moved on to Python. They have problem sets and tutorials - and you'll be graded based on your submission, which feels a lot like professional education.

Starthere.fm is a podcast that's amazing for motivation, tips, etc for a web developer that's starting out. Dain is very real and honest about what to do, and the tips he gives in the podcasts are extremely helpful.